+++ pretext/constants.py
@@ -142,6 +142,7 @@ ASSET_TO_XPATH = {
     "myopenmath": ".//myopenmath",
     "dynamic-subs": ".//statement[.//fillin and ancestor::exercise/evaluation]",
     "references": ".//biblio",
+    "stack": ".//stack",
 }
 ASSETS = ["ALL"] + list(ASSET_TO_XPATH.keys())
 
@@ -159,6 +160,8 @@ ASSET_TO_DIR = {
     "mermaid": ["mermaid"],
     "myopenmath": ["problems"],
     "dynamic-subs": ["dynamic_subs"],
+    "references": ["references"],
+    "stack": ["stack"],
 }
 
 ASSET_FORMATS: t.Dict[str, t.Dict[str, t.List[str]]] = {